[#65993] CMake: Always build as shared library
Object library is used to build all the files but skip linking. It was
used to define new library in Renode using tlib objects with additional
sources.

It turned out, however, that modifying tlib target after adding tlib
subdirecory is much simpler so skipping linking isn't really needed.
Especially that it made tlib impossible to build on its own on OSs other
than Linux.

Besides, the `TLIB_LIB_TYPE` wasn't properly set and there was only an
impression that it was used on Linux.