[#65993] CMake: Add `TLIB_PROFILING_BUILD` definitions only once
TCG and SF3 are added as subdirectories which, compared to using
ExternalProject, applies `add_definitions` definitions from before
they're added. Therefore we don't need to set `omit-frame-pointer` etc.
yet again in both of them.

Also, SF3 is a submodule and it's a separate project so using `TLIB_` in
the variable name was problematic. We can easily add the definitions in
tlib's CMake though so let's move `INLINE_LEVEL` setting there.